add second module content
